requirements.txt
Babel==1.3
BeautifulSoup4==4.4.1
cmsplugin_filer==1.0.1
cmsplugin-plaintext==0.1.1a2
Django==1.8.8
django-appconf==1.0.1
django-classy-tags==0.7.1
django-cms==3.2.1
django-compressor==2.0
django-extra-views==0.6.4
django-filer==1.1.1
django-formtools==1.0
django-haystack==2.3.2
django-mptt==0.8.0
django-oscar==1.1.1
django-phonenumber-field==1.0.0
django-polymorphic==0.8.1
django-reversion==1.10.1
django-sekizai==0.9.0
django-tables2==1.0.7
django-treebeard==4.0
django-widget-tweaks==1.3
djangocms-admin-style==1.0.9
djangocms-cascade==0.7.1
djangocms-text-ckeditor==2.8.1
easy-thumbnails==2.3
easywebdav==1.2.0
fake-factory==0.5.3
git+https://github.com/mi6gan/djnagocms_plugin_bootstrap_columns
gunicorn
html5lib==0.9999999
jsonfield==1.0.3
mock==1.0.1
phonenumbers==6.3.0
Pillow==2.7.0
psycopg2==2.6.1
purl==1.2
pycountry==1.20
pytz==2015.7
rcssmin==1.0.6
requests==2.9.1
rjsmin==1.0.12
six==1.10.0
sorl-thumbnail==12.2
unicodecsv==0.14.1
Unidecode==0.4.19
